Moved the singularity tests along with the cvmfs tests
