+++ files/test_sequence
@@ -42,7 +42,6 @@ test_53_gums
 test_54_gratia
 test_55_condorce
 test_59_gsiopenssh
-test_60_singularity
 test_74_slurm
 test_75_gums
 test_76_tomcat
