+++ osgtest/tests/test_38_cvmfs.py
@@ -7,7 +7,19 @@ import tempfile
 class TestCvmfs(osgunittest.OSGTestCase):
 
     __check_path = '/cvmfs/cms.cern.ch/cmsset_default.sh'
+    __cvmfs_image = '/cvmfs/singularity.opensciencegrid.org/opensciencegrid/osg-wn:3.3-el6'
     core.config['cvmfs.debug-dirs'] = []
+
+    def mountSingularityCVMFSRepo(self, repo):
+        command = ('mkdir', '-p', /cvmfs + repo)
+        status, stdout, stderr = core.system(command, False)
+        if status != 0:
+            self.fail("failed to mkdir /cvmfs/%s" % repo)
+
+        command = ('mount', '-t', 'cvmfs', 'repo', '/cvmfs/' + repo)
+        status, stdout, stderr = core.system(command, False)
+        if status != 0:
+            self.fail("failed to mount: %s, error: %s" % (repo,stdout))
 
     def debug_cvmfs(self, repo):
         temp_dir = tempfile.mkdtemp()
@@ -97,3 +109,32 @@ class TestCvmfs(osgunittest.OSGTestCase):
         # If the previous command failed, output better debug info
         if status != 0:
             self.debug_cvmfs(oasis_repo)
+
+    def test_04_singularity(self):
+        core.skip_ok_unless_installed('singularity-runtime')
+        core.skip_ok_unless_installed('cvmfs')
+        singularity_repo = 'singularity.opensciencegrid.org'
+        if core.el_release() <= 6:
+             self.mountSingularityCVMFSRepo(singularity_repo)
+        core.state['cvmfs.mounted'] = False
+
+        command = ('ls', '/cvmfs')
+        status, stdout, stderr = core.system(command, False)
+        file_exists = os.path.exists('/cvmfs')
+        self.assert_(file_exists, 'Cvmfs mount point missing')
+        core.state['cvmfs.mounted'] = True
+
+        command = ('ls', '/cvmfs/' + singularity_repo)
+        status, stdout, stderr = core.system(command, False)
+
+        # If the previous command failed, output better debug info                                                                                                                                 
+        if status != 0:
+            self.fail("failed to find /cvmfs/%s" % singularity_repo)
+
+        command = ('ls', self.__cvmfs_image)
+        status, stdout, stderr = core.system(command, False)
+        self.assert_(file_exists, 'cvfms image missing')
+        command= ('singularity', 'exec', '--bind', '/cvmfs', self.__cvmfs_image, 'echo', 'working singularity image')
+        status, stdout, stderr = core.system(command, False)
+        fail = core.diagnose('singularity checking a file', command, status, stdout, stderr)
+        self.assertEqual(status, 0, fail)
